+++ lib/tbot/config/configtemplate_kubernetes_test.go
@@ -22,92 +22,11 @@ import (
 	"os"
 	"path/filepath"
 	"testing"
-	"time"
 
-	"github.com/gravitational/teleport/api/client/proto"
-	"github.com/gravitational/teleport/api/constants"
-	"github.com/gravitational/teleport/lib/auth/native"
-	"github.com/gravitational/teleport/lib/auth/testauthority"
-	"github.com/gravitational/teleport/lib/fixtures"
-	"github.com/gravitational/teleport/lib/services"
 	"github.com/gravitational/teleport/lib/tbot/botfs"
-	"github.com/gravitational/teleport/lib/tbot/identity"
-	"github.com/gravitational/teleport/lib/tlsca"
 	"github.com/gravitational/teleport/lib/utils/golden"
-	"github.com/jonboulle/clockwork"
 	"github.com/stretchr/testify/require"
-	"golang.org/x/crypto/ssh"
 )
-
-// getTestIdent returns a mostly-valid bot Identity without starting up an
-// entire Teleport server instance.
-func getTestIdent(t *testing.T, username string, k8sCluster string) *identity.Identity {
-	ca, err := tlsca.FromKeys([]byte(fixtures.TLSCACertPEM), []byte(fixtures.TLSCAKeyPEM))
-	require.NoError(t, err)
-
-	privateKey, sshPublicKey, err := native.GenerateKeyPair()
-	require.NoError(t, err)
-
-	sshPrivateKey, err := ssh.ParseRawPrivateKey(privateKey)
-	require.NoError(t, err)
-
-	tlsPublicKeyPEM, err := tlsca.MarshalPublicKeyFromPrivateKeyPEM(sshPrivateKey)
-	require.NoError(t, err)
-
-	tlsPublicKey, err := tlsca.ParsePublicKeyPEM(tlsPublicKeyPEM)
-	require.NoError(t, err)
-
-	// Note: it'd be nice to make this more universally useful in our tests at
-	// some point.
-	clock := clockwork.NewFakeClock()
-	notAfter := clock.Now().Add(time.Hour)
-	id := tlsca.Identity{
-		Username:          username,
-		KubernetesUsers:   []string{"foo"},
-		KubernetesGroups:  []string{"bar"},
-		RouteToCluster:    mockClusterName,
-		KubernetesCluster: k8sCluster,
-	}
-	subject, err := id.Subject()
-	require.NoError(t, err)
-	certBytes, err := ca.GenerateCertificate(tlsca.CertificateRequest{
-		Clock:     clock,
-		PublicKey: tlsPublicKey,
-		Subject:   subject,
-		NotAfter:  notAfter,
-	})
-	require.NoError(t, err)
-
-	caSigner, err := ssh.ParsePrivateKey([]byte(fixtures.SSHCAPrivateKey))
-	require.NoError(t, err)
-	ta := testauthority.New()
-	sshCertBytes, err := ta.GenerateUserCert(services.UserCertParams{
-		CASigner:          caSigner,
-		PublicUserKey:     sshPublicKey,
-		Username:          username,
-		CertificateFormat: constants.CertificateFormatStandard,
-		TTL:               time.Minute,
-		AllowedLogins:     []string{"foo"},
-		RouteToCluster:    mockClusterName,
-	})
-
-	require.NoError(t, err)
-
-	certs := &proto.Certs{
-		SSH:        sshCertBytes,
-		TLS:        certBytes,
-		TLSCACerts: [][]byte{[]byte(fixtures.TLSCACertPEM)},
-		SSHCACerts: [][]byte{[]byte(fixtures.SSHCAPublicKey)},
-	}
-
-	ident, err := identity.ReadIdentityFromStore(&identity.LoadIdentityParams{
-		PrivateKeyBytes: privateKey,
-		PublicKeyBytes:  tlsPublicKeyPEM,
-	}, certs, identity.DestinationKinds()...)
-	require.NoError(t, err)
-
-	return ident
-}
 
 // TestTemplateKubernetesRender renders a Kubernetes template and compares it
 // to the saved golden result.
@@ -140,7 +59,7 @@ func TestTemplateKubernetesRender(t *testing.T) {
 		},
 	}
 
-	ident := getTestIdent(t, "bot-test", k8sCluster)
+	ident := getTestIdent(t, "bot-test", kubernetesRequest(k8sCluster))
 
 	err = template.Render(context.Background(), mockBot, ident, dest)
 	require.NoError(t, err)
