+++ lib/tbot/config/configtemplate_ssh_host_cert.go
@@ -33,7 +33,7 @@ const (
 	defaultSSHHostCertPrefix = "ssh_host"
 )
 
-// TemplateSSHClient contains parameters for the ssh_config config
+// TemplateSSHHostCert contains parameters for the ssh_config config
 // template
 type TemplateSSHHostCert struct {
 	// Prefix is the filename prefix for the generated SSH host
@@ -44,6 +44,7 @@ type TemplateSSHHostCert struct {
 	Principals []string `yaml:"principals"`
 }
 
+// CheckAndSetDefaults validates a TemplateSSHHostCert.
 func (c *TemplateSSHHostCert) CheckAndSetDefaults() error {
 	if c.Prefix == "" {
 		c.Prefix = defaultSSHHostCertPrefix
@@ -56,10 +57,12 @@ func (c *TemplateSSHHostCert) CheckAndSetDefaults() error {
 	return nil
 }
 
+// Name returns the name for the ssh_host_cert template.
 func (c *TemplateSSHHostCert) Name() string {
 	return TemplateSSHHostCertName
 }
 
+// Describe lists the files to be generated by the ssh_host_cert template.
 func (c *TemplateSSHHostCert) Describe(destination bot.Destination) []FileDescription {
 	ret := []FileDescription{
 		{
@@ -73,6 +76,7 @@ func (c *TemplateSSHHostCert) Describe(destination bot.Destination) []FileDescri
 	return ret
 }
 
+// Render generates SSH host cert files.
 func (c *TemplateSSHHostCert) Render(ctx context.Context, bot Bot, currentIdentity *identity.Identity, destination *DestinationConfig) error {
 	dest, err := destination.GetDestination()
 	if err != nil {
@@ -80,8 +84,7 @@ func (c *TemplateSSHHostCert) Render(ctx context.Context, bot Bot, currentIdenti
 	}
 
 	// We'll need a client for the impersonated identity to request the certs.
-	botCfg := bot.Config()
-	authClient, err := bot.AuthenticatedUserClientFromIdentity(ctx, currentIdentity, botCfg.AuthServer)
+	authClient, err := bot.AuthenticatedUserClientFromIdentity(ctx, currentIdentity)
 	if err != nil {
 		return trace.Wrap(err)
 	}
@@ -100,6 +103,7 @@ func (c *TemplateSSHHostCert) Render(ctx context.Context, bot Bot, currentIdenti
 
 	// For now, we'll reuse the bot's regular TTL, and hostID and nodeName are
 	// left unset.
+	botCfg := bot.Config()
 	key.Cert, err = authClient.GenerateHostCert(key.MarshalSSHPublicKey(),
 		"", "", c.Principals,
 		clusterName, types.RoleNode, botCfg.CertificateTTL)
